bug fix in course CRUD :bug:
adjustment of the components/Section/Cursos/Cursos.vue component to replace the use of campusBaseQueryKey with campusBaseQueryKey, ensuring consistency with query keys used in other parts of the application. The problem occurred because query invalidation was incorrectly affecting the Courses grid, forcing information to be updated based on an incorrect key. With this correction, the component now correctly uses the campusBaseQueryKey key, aligning with expected behavior and avoiding inconsistencies in data loading.