Feat: New Calendar Functions
Now, in the school calendar, specifically in the calendar component, there is now the option to switch between months, as long as they are from the same year. Partial integration with the API was also done, to avoid future integration problems, and also to explore possibilities for new features and improvements. A change was also made to the types in the calendar session, to better adapt to the API. Some advances were also made in the events component, which is now also partially linked to the API.

~ More features will be available soon.